requirements.txt
accelerate==1.3.0
accessible-pygments==0.0.5
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
arxiv==1.4.3
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
blis==1.2.0
bs4==0.0.1
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
cmudict==1.0.32
commonmark==0.9.1
confection==0.1.5
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.11
demucs==4.0.1
distro==1.9.0
docutils==0.21.2
dora_search==0.1.12
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
Flask==2.0.1
Flask-Cors==3.0.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gradio==3.20.1
grobid-client-python==0.0.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
itchat==1.2.32
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
lameenc==1.8.1
langcodes==3.5.0
language_data==1.3.0
linkify-it-py==2.0.3
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
narwhals==1.27.1
networkx==3.4.2
npm==0.1.1
numpy==2.2.2
ollama==0.4.7
omegaconf==2.3.0
openai==0.27.8
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
openunmix==1.3.0
optional-django==0.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
Pillow==9.4.0
pptx2md==2.0.6
preshed==3.0.9
propcache==0.2.1
psutil==6.1.1
pybase64==1.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydata-sphinx-theme==0.15.4
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.22.3
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyphen==0.17.2
pypinyin==0.53.0
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.1
recommonmark==0.7.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.31.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.2
scikit-learn==1.6.1
scipdf-parser==0.52
scipy==1.15.1
sentence-transformers==3.4.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.1.3
sphinx-book-theme==1.1.4
sphinx-markdown-tables==0.0.17
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxemoji==0.3.1
srsly==2.5.1
starlette==0.45.3
submitit==1.5.2
sympy==1.13.1
tenacity==8.2.2
termcolor==2.5.0
textstat==0.7.5
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.4.0
timm==1.0.14
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.1
treetable==0.2.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
urllib3==1.25.11
uvicorn==0.34.0
volcengine-python-sdk==1.0.124
wasabi==1.1.3
weasel==0.4.1
websockets==15.0
Werkzeug==2.0.1
wrapt==1.17.2
wxpy==0.3.9.8
xlrd==2.0.1
XlsxWriter==3.2.2
yarl==1.18.3
zipp==3.21.0
